requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
docx2txt
PyPDF2
langchain_google_genai